+++ game/game_libs/ui/controls/PlayerModelView.cpp
@@ -73,7 +73,7 @@ void CMenuPlayerModelView::VidInit()
 	ent->latched.prevcontroller[1] = 127;
 	ent->latched.prevcontroller[2] = 127;
 	ent->latched.prevcontroller[3] = 127;
-	ent->origin[0] = ent->curstate.origin[0] = 45.0f / tan(DEG2RAD(refdef.fov_y / 2.0f));
+	ent->origin[0] = ent->curstate.origin[0] = 45.0f / tanf(DEG2RAD(refdef.fov_y / 2.0f));
 	ent->origin[2] = ent->curstate.origin[2] = 2.0f;
 	ent->angles[1] = ent->curstate.angles[1] = 180.0f;
 
@@ -156,7 +156,7 @@ void CMenuPlayerModelView::Draw()
 		if ( uiStatic.enableAlphaFactor )
 		{
 			ent->curstate.rendermode = kRenderTransTexture;
-			ent->curstate.renderamt = uiStatic.alphaFactor * 255;
+			ent->curstate.renderamt = static_cast<int>(uiStatic.alphaFactor * 255);
 		}
 		else
 		{
@@ -166,7 +166,8 @@ void CMenuPlayerModelView::Draw()
 
 		if ( mouseYawControl )
 		{
-			float diffX = uiStatic.cursorX - prevCursorX;
+			float diffX = static_cast<float>(uiStatic.cursorX - prevCursorX);
+
 			if ( diffX )
 			{
 				float yaw = ent->angles[1];
@@ -216,7 +217,7 @@ assume refdef is valid
 */
 void CMenuPlayerModelView::CalcFov()
 {
-	float x = refdef.viewport[2] / tan(DEG2RAD(refdef.fov_x) * 0.5f);
-	float half_fov_y = atan(refdef.viewport[3] / x);
+	float x = refdef.viewport[2] / tanf(DEG2RAD(refdef.fov_x) * 0.5f);
+	float half_fov_y = atanf(refdef.viewport[3] / x);
 	refdef.fov_y = RAD2DEG(half_fov_y) * 2;
 }
